requirements.txt
Flask>=3.0,<4
gunicorn>=22,<24
Pillow>=10,<11
